+++ src/Calculators/Dominion/MilitaryCalculator.php
@@ -155,7 +155,6 @@ class MilitaryCalculator
         $spellHowling = 10;
         $spellKillingRage = 10;
         $spellNightfall = 5;
-        $spellWarsong = 10;
 
         // Gryphon Nests
         if ($dominion->calc !== null && !isset($dominion->calc['invasion'])) {
@@ -214,10 +213,6 @@ class MilitaryCalculator
             if (isset($dominion->calc['nightfall'])) {
                 $multiplier += ($spellNightfall / 100);
             }
-
-            if (isset($dominion->calc['warsong'])) {
-                $multiplier += ($spellWarsong / 100);
-            }
         } else {
             $multiplier += $this->spellCalculator->getActiveSpellMultiplierBonus($dominion, [
                 'bloodrage' => $spellBloodrage,
@@ -225,7 +220,6 @@ class MilitaryCalculator
                 'howling' => $spellHowling,
                 'killing_rage' => $spellKillingRage,
                 'nightfall' => $spellNightfall,
-                'warsong' => $spellWarsong,
             ]);
         }
 
@@ -525,10 +519,8 @@ class MilitaryCalculator
             $unitPower += $this->getUnitPowerFromStaggeredLandRangePerk($dominion, $landRatio, $unit, $powerType);
         }
 
-        if ($target !== null || ($dominion->calc !== null && !isset($dominion->calc['invasion']))) {
-            $unitPower += $this->getUnitPowerFromVersusRacePerk($dominion, $target, $unit, $powerType);
-            $unitPower += $this->getUnitPowerFromVersusBuildingPerk($dominion, $target, $unit, $powerType);
-        }
+        $unitPower += $this->getUnitPowerFromVersusRacePerk($dominion, $target, $unit, $powerType);
+        $unitPower += $this->getUnitPowerFromVersusBuildingPerk($dominion, $target, $unit, $powerType);
 
         return $unitPower;
     }
@@ -718,7 +710,7 @@ class MilitaryCalculator
 
     protected function getUnitPowerFromVersusBuildingPerk(Dominion $dominion, Dominion $target = null, Unit $unit, string $powerType): float
     {
-        if ($target === null && $dominion->calc == null) {
+        if ($target === null && $dominion->calc === null) {
             return 0;
         }
 
@@ -810,6 +802,16 @@ class MilitaryCalculator
     {
         $multiplier = 1;
 
+        // Values (percentages)
+        $forestHavenBonus = 2;
+        $forestHavenBonusMax = 20;
+
+        // Forest Havens
+        $multiplier += min(
+            (($dominion->building_forest_haven / $this->landCalculator->getTotalLand($dominion)) * $forestHavenBonus),
+            ($forestHavenBonusMax / 100)
+        );
+
         // Racial bonus
         $multiplier += $dominion->race->getPerkMultiplier('spy_strength');
 
@@ -882,6 +884,16 @@ class MilitaryCalculator
     {
         $multiplier = 1;
 
+        // Values (percentages)
+        $wizardGuildBonus = 2;
+        $wizardGuildBonusMax = 20;
+
+        // Wizard Guilds
+        $multiplier += min(
+            (($dominion->building_wizard_guild / $this->landCalculator->getTotalLand($dominion)) * $wizardGuildBonus),
+            ($wizardGuildBonusMax / 100)
+        );
+
         // Racial bonus
         $multiplier += $dominion->race->getPerkMultiplier('wizard_strength');
 
@@ -921,7 +933,7 @@ class MilitaryCalculator
         // Docks
         $boatsProtected = static::BOATS_PROTECTED_PER_DOCK * $dominion->building_dock;
         // Habor
-        $boatsProtected *= (1 + $this->improvementCalculator->getImprovementMultiplierBonus($dominion, 'harbor'));
+        $boatsProtected *= (1 + $this->improvementCalculator->getImprovementMultiplierBonus($dominion, 'harbor') * 2);
         return $boatsProtected;
     }