+++ src/Calculators/Dominion/SpellCalculator.php
@@ -44,9 +44,9 @@ class SpellCalculator
         $spellInfo = $this->spellHelper->getSpellInfo($spell, $dominion->race);
         $totalLand = $this->landCalculator->getTotalLand($dominion);
 
-        // Cost reduction from wizard guilds (2x ratio, max 40%)
+        // Cost reduction from wizard guilds (3x ratio, max 30%)
         $wizardGuildRatio = ($dominion->building_wizard_guild / $totalLand);
-        $spellCostMultiplier = (1 - clamp(2 * $wizardGuildRatio, 0, 0.4));
+        $spellCostMultiplier = (1 - clamp(3 * $wizardGuildRatio, 0, 0.3));
         $spellCostMultiplier += $dominion->getTechPerkMultiplier('spell_cost');
 
         return round($spellInfo['mana_cost'] * $totalLand * $spellCostMultiplier);