+++ src/Console/Commands/Game/TickCommand.php
@@ -32,14 +32,8 @@ class TickCommand extends Command implements CommandInterface
      */
     public function handle(): void
     {
-        if (now()->hour === 0) {
-            $this->tickService->tickDaily();
-        }
-
+        $this->tickService->tickDaily();
         $this->tickService->tickHourly();
-
-        if (now()->hour === 0) {
-            $this->tickService->updateDailyRankings();
-        }
+        $this->tickService->updateDailyRankings();
     }
 }